# account-service/src/main/java/com/bank/account_service/dto/MovementRequestDto.java
package com.bank.account_service.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Positive;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class MovementRequestDto {

    @NotNull(message = "Account ID is required")
    private Long accountId;

    @NotBlank(message = "Movement type is required (DEBIT or CREDIT)")
    private String movementType;

    @NotNull(message = "Amount is required")
    @Positive(message = "Amount must be greater than zero")
    private BigDecimal amount;
}